netconf-poller/requirements.txt
ncclient
pyyaml
zabbix-api
